KVM: x86: Untag address in instruction emulator
Call get_untagged_addr() interface in __linearize() in instruction emulator
for 64-bit mode.

Vendor specific code has the flexibility to decide to do address untag or
not according to the 'flags' passed in.

Note LAM doesn't apply to loadings of memory management registers and base
address of segment registers, the values still need to be canonical. So no
calls of get_untagged_addr() added for implicit system access for now.

Co-developed-by: Robert Hoo <robert.hu@linux.intel.com>
Signed-off-by: Robert Hoo <robert.hu@linux.intel.com>
Signed-off-by: Binbin Wu <binbin.wu@linux.intel.com>
Reviewed-by: Chao Gao <chao.gao@intel.com>